move duplicate constant to user models
