--- src/modules/user/pages/register/constants.ts
@@ -1,6 +1,0 @@
-import { FormProps } from 'antd/lib/form';
-
-export const layout: FormProps = {
-  labelCol: { span: 8 },
-  wrapperCol: { span: 24 },
-};
